+++ lib/types/types/common.d.ts
@@ -1,0 +1,2 @@
+export type Formatter<T, P> = (payload: P) => T;
+export type BaseAttr<T, P> = T | Formatter<T, P>;