+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * border-color -> #dadce0
     * li:hover background-color -> #e8e8e9, .menu-item-tip: color -> #000
     * divide line: color -> ,margin-left/right -> 1px
+* change: ts `declarationDir` lib/index.d.ts -> lib/types/index.d.ts
 ## v1.4.4
 * remove: package.json `source`.
 * optimize: browser rerender cause by `utils -> getWindowSize()`.